here the admin credentials has been changed

+++ backend/database/migrations/2025_11_03_160100_create_order_status_updates_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
                 $table->unsignedBigInteger('order_id');
                 $table->string('from_status')->nullable();
                 $table->string('to_status');
-                $table->unsignedBigInteger('changed_by_admin_id')->nullable();
+                $table->unsignedInteger('changed_by_admin_id')->nullable();
                 $table->text('note')->nullable();
                 $table->timestamps();
